Renaming Example module dir to List, fixing links in list, cleaning up list component
