+++ src/main.js
@@ -4,8 +4,8 @@ import Router from 'cerebral-module-router';
 import Controller from 'cerebral';
 import Model from 'cerebral-model-baobab';
 import {Container} from 'cerebral-view-react';
-import List from './modules/Example';
-import ItemList from './modules/Example/components/ItemList';
+import List from './modules/List';
+import ItemList from './modules/List/components/ItemList';
 import Http from 'cerebral-module-http';
 
 const controller = Controller(Model({}));
