+++ src/modules/List/signals/categoryChanged.js
